+++ react/logisticsInfo.js
@@ -6,6 +6,7 @@ import {
 } from '@vtex/delivery-packages/dist/delivery-channel'
 import { helpers } from 'vtex.address-form'
 import {
+  findSlaOption,
   isFromCurrentSeller,
   hasCurrentDeliveryChannel,
   findSlaWithChannel,
@@ -142,12 +143,15 @@ export function getNewLogisticsInfoIfDelivery({
     }
   }
 
+  const selectedSla = findSlaOption([logisticsInfo], logisticsInfo.selectedSla)
+  const shouldUseSelectedSla = selectedSla && selectedSla.deliveryChannel === channel
+
   if (hasCurrentDeliveryChannel(logisticsInfo, channel)) {
     return {
       ...logisticsInfo,
       addressId: actionAddress.addressId,
       selectedDeliveryChannel: channel,
-      selectedSla: logisticsInfo.selectedSla || defaultSlaSelection.id,
+      selectedSla: shouldUseSelectedSla ? logisticsInfo.selectedSla : defaultSlaSelection.id,
     }
   }
 }