+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Pickup SLAs can not be selected when the `selectedDeliveryChannel` is `delivery`.
+
 ## [0.2.9] - 2020-04-30
 
 ### Fixed